feat: extract preview document logic into its own store
